♻️ refactor: add trpc query client with react-query (#3282)
* ♻️ refactor: refactor the trpc client with react query mode

* ♻️ refactor: refactor the trpc client with react query mode
