+++ package.json
@@ -121,8 +121,10 @@
     "@next/third-parties": "^14.2.4",
     "@sentry/nextjs": "^7.118.0",
     "@t3-oss/env-nextjs": "^0.10.1",
+    "@tanstack/react-query": "^5.51.11",
     "@trpc/client": "next",
     "@trpc/next": "next",
+    "@trpc/react-query": "next",
     "@trpc/server": "next",
     "@vercel/analytics": "^1.3.1",
     "@vercel/speed-insights": "^1.0.12",
